Update selectors, more listeners, and fix cases.
There was a problem causing only one selection to be read. There
was also a problem causing sensitivity in the computer selection
to be read in upper case causing for the play to always lose.

By adding more selectors, it allowed for more than one button
to be activated and sent to playRound function. Adding more
event listeners was required to fix this issue. Also changed
computer selection for solving case sensitivity issue.